fix: more robust handling of custom shortcuts
* user was able to set a nextWindowShortcut which contained modifiers from holdShortcut
* detection of conflicts was incorrect in some cases, mostly because holdShortcut and nextWindowShortcut should be treated as 1 shortcut
* after the user forces a shortcut to apply, resetting the existing one with the same keys, it wouldn't save to UserDefaults
* forcing the holdShortcut to replace another holdShortcut would reset it, even though holdShortcut should always be set with some value

+++ src/logic/events/KeyboardEvents.swift
@@ -39,8 +39,8 @@ class KeyboardEvents {
         if shortcut.keyCode != .none {
             let id = globalShortcutsIds[controlId]!
             let hotkeyId = EventHotKeyID(signature: signature, id: UInt32(id))
-            let key = UInt32(shortcut.carbonKeyCode)
-            let mods = UInt32(shortcut.carbonModifierFlags)
+            let key = shortcut.carbonKeyCode
+            let mods = shortcut.carbonModifierFlags
             let options = UInt32(kEventHotKeyNoOptions)
             var shortcutsReference: EventHotKeyRef?
             RegisterEventHotKey(key, mods, hotkeyId, shortcutEventTarget, options, &shortcutsReference)
